+++ tests/pools/cluster/dynamic.test.js
@@ -21,14 +21,14 @@ describe('Dynamic cluster pool test suite', () => {
 
   it('Verify that new workers are created when required, max size is not exceeded and that after a while new workers will die', async () => {
     const promises = []
-    let busy = 0
-    pool.emitter.on('busy', () => busy++)
+    let poolBusy = 0
+    pool.emitter.on('busy', () => poolBusy++)
     for (let i = 0; i < max * 2; i++) {
       promises.push(pool.execute({ test: 'test' }))
     }
     expect(pool.workers.length).toBeLessThanOrEqual(max)
     expect(pool.workers.length).toBeGreaterThan(min)
-    expect(busy > 1).toBeTruthy()
+    expect(poolBusy).toEqual(max + 1)
     const numberOfExitEvents = await TestUtils.waitExits(pool, max - min)
     expect(numberOfExitEvents).toBe(max - min)
   })